+++ src/UnknownType.php
@@ -16,24 +16,24 @@ final class UnknownType extends Type
         return true;
     }
 
+    public function name(): string
+    {
+        return 'unknown type';
+    }
+
     public function asString(): string
     {
         return '';
     }
 
-    public function asReturnTypeDeclaration(): string
-    {
-        return $this->asString();
-    }
-
     /**
-     * @deprecated Use asReturnTypeDeclaration() instead
+     * @deprecated
      *
      * @codeCoverageIgnore
      */
     public function getReturnTypeDeclaration(): string
     {
-        return $this->asReturnTypeDeclaration();
+        return '';
     }
 
     public function allowsNull(): bool