+++ tests/unit/GenericObjectTypeTest.php
@@ -38,7 +38,7 @@ final class GenericObjectTypeTest extends TestCase
 
     public function testCanBeRepresentedAsStringForReturnTypeDeclaration(): void
     {
-        $this->assertEquals(': object', $this->type->asReturnTypeDeclaration());
+        $this->assertEquals(': object', $this->type->getReturnTypeDeclaration());
     }
 
     public function testMayAllowNull(): void
@@ -52,7 +52,7 @@ final class GenericObjectTypeTest extends TestCase
     {
         $type = new GenericObjectType(true);
 
-        $this->assertEquals(': ?object', $type->asReturnTypeDeclaration());
+        $this->assertEquals(': ?object', $type->getReturnTypeDeclaration());
     }
 
     public function testObjectCanBeAssignedToGenericObject(): void